Set complete=false in redeemmultisigout on invalid sigs
If the raw transaction provided by the caller contains invalid signatures, set
complete=false to indicate that the transaction is not ready to send.

This situation (where the user has manually modified the transaction after a
previous redeemmultisigout call, which may add some of the required
signatures) is unfortunate and there are possibly better ways we could deal
with this, such as removing the invalid signature, or returning a more
descriptive error for the failed script evaluation.